[refactor] RTK의 hook안에 넘기는 option 분리로인한 수정
